Move all code to unit_test_framework.h.
